ルートパスアクセス時にHome/page配下のHomeを描画するように修正
