+++ src/app/types/Todo.ts
@@ -1,0 +1,5 @@
+export type Todo = {
+  id: number
+  title: string
+  text?: string
+}
